+++ src/iOS.Core/Renderers/CustomLabelRenderer.cs
@@ -1,19 +1,19 @@
 ﻿using System.ComponentModel;
-using Bit.iOS.Renderers;
-using Bit.iOS.Utilities;
+using Bit.iOS.Core.Renderers;
+using Bit.iOS.Core.Utilities;
 using UIKit;
 using Xamarin.Forms;
 using Xamarin.Forms.Platform.iOS;
 
-[assembly: ExportRenderer(typeof(Button), typeof(CustomButtonRenderer))]
-namespace Bit.iOS.Renderers
+[assembly: ExportRenderer(typeof(Label), typeof(CustomLabelRenderer))]
+namespace Bit.iOS.Core.Renderers
 {
-    public class CustomButtonRenderer : ButtonRenderer
+    public class CustomLabelRenderer : LabelRenderer
     {
-        protected override void OnElementChanged(ElementChangedEventArgs<Button> e)
+        protected override void OnElementChanged(ElementChangedEventArgs<Label> e)
         {
             base.OnElementChanged(e);
-            if (Control != null && e.NewElement is Button)
+            if (Control != null && e.NewElement is Label)
             {
                 UpdateFont();
             }
@@ -22,7 +22,9 @@ namespace Bit.iOS.Renderers
         protected override void OnElementPropertyChanged(object sender, PropertyChangedEventArgs e)
         {
             base.OnElementPropertyChanged(sender, e);
-            if (e.PropertyName == Button.FontProperty.PropertyName)
+            if (e.PropertyName == Label.FontProperty.PropertyName ||
+                e.PropertyName == Label.TextProperty.PropertyName ||
+                e.PropertyName == Label.FormattedTextProperty.PropertyName)
             {
                 UpdateFont();
             }
@@ -30,7 +32,7 @@ namespace Bit.iOS.Renderers
 
         private void UpdateFont()
         {
-            var pointSize = iOSHelpers.GetAccessibleFont<Button>(Element.FontSize);
+            var pointSize = iOSHelpers.GetAccessibleFont<Label>(Element.FontSize);
             if (pointSize != null)
             {
                 Control.Font = UIFont.FromDescriptor(Element.Font.ToUIFont().FontDescriptor, pointSize.Value);
