+++ src/iOS.Core/Renderers/CustomPickerRenderer.cs
@@ -1,11 +1,11 @@
-﻿using Bit.iOS.Renderers;
-using Bit.iOS.Utilities;
+﻿using Bit.iOS.Core.Renderers;
+using Bit.iOS.Core.Utilities;
 using UIKit;
 using Xamarin.Forms;
 using Xamarin.Forms.Platform.iOS;
 
 [assembly: ExportRenderer(typeof(Picker), typeof(CustomPickerRenderer))]
-namespace Bit.iOS.Renderers
+namespace Bit.iOS.Core.Renderers
 {
     public class CustomPickerRenderer : PickerRenderer
     {
@@ -23,7 +23,7 @@ namespace Bit.iOS.Renderers
 
         private void UpdateKeyboardAppearance()
         {
-            if (!Core.Utilities.ThemeHelpers.LightTheme)
+            if (!ThemeHelpers.LightTheme)
             {
                 Control.KeyboardAppearance = UIKeyboardAppearance.Dark;
             }
