+++ src/iOS.Core/Renderers/CustomViewCellRenderer.cs
@@ -1,10 +1,10 @@
-﻿using Bit.iOS.Renderers;
+﻿using Bit.iOS.Core.Renderers;
 using UIKit;
 using Xamarin.Forms;
 using Xamarin.Forms.Platform.iOS;
 
 [assembly: ExportRenderer(typeof(ViewCell), typeof(CustomViewCellRenderer))]
-namespace Bit.iOS.Renderers
+namespace Bit.iOS.Core.Renderers
 {
     public class CustomViewCellRenderer : ViewCellRenderer
     {
@@ -12,7 +12,7 @@ namespace Bit.iOS.Renderers
 
         public CustomViewCellRenderer()
         {
-            _noSelectionStyle = (Color)Xamarin.Forms.Application.Current.Resources["BackgroundColor"] != Color.White;
+            _noSelectionStyle = (Color)Application.Current.Resources["BackgroundColor"] != Color.White;
         }
 
         public override UITableViewCell GetCell(Cell item, UITableViewCell reusableCell, UITableView tv)
